Add initial dotfiles configuration
This commit adds initial base dotfiles configuration that was used in
late February 2024.

This version of config was a continuous work since around summer of
2020, when powerlevel9k was depricated and my config was updated to
start using powerlevel10k. With this improvement additional work to get
rid of most vim configuration and start using minimalistic vim setup was
attempted, which lead to vim abandonment in the long run and limiting
terminal usage for a couple of years heavily.

This changes and improvements were mostly undocumented and this commit
aims just to preserve them before major configuration changes are going
to be introduced.

My plan is to go from there and modernise and standardise my stack going
forward, mainly:
- I want to use system agnostic terminal setup, therefore:
- I'm planning to change terminal emulator to something that can be used
  on Mac OS, Linux and even Windows as i have recently started
  experimenting with WSL and webdev on my gaming machine, something like
  Alacritty might be a good and modern solution
- I will most likely continue to run on zsh, even though more devops
  work that I would like to take up tempts me to go back to bash as
  alpine containers usually come with bash as a default
- I'd like to get rid of powerlevel as it is quite heavy weight, I'm
  considering some simple theme replacement, or experimenting with more
  modern soultions such as oh-my-posh
- I'd love to go back to vim, I'm considering giving neovim a try, maybe
  a preconfigured setup one like Lazyvim would be a good choice to start
  with

This snapshot of configuration marks an end of a very long, over a
decade journey that was based on my personal config going from bash to
zsh, starting and stopping using vim and using powerlevel as a theme.
Main goal for now is to modernise some things, secure potential usage
on other operating systems, not only mac OS, and to be honest just bring
some change to my toolset as I fell like I currently keep a lot of old
habits that keep me from personal growth as a developer and engineer.