Change used terminal to Alacritty
Alacritty now replaces iTerm2 as a main terminal emulator

This change happens after around 15 years of using iTerm2 as a terminal
emulator that I have used on a daily basis

Alacritty has a couple of features that make sense to me and that
position it as a more modern and better emulator than iTerm2:
- it is cross-platform, so I can freely use it on Linux and Windows -
  for many years that has been a problem as I had a barrier of using
  other OSes after many years of working on mac OS, as my setup was very
  mac OS centric and hard to port
- it is GPU accelerated which makes it more snappy in daily use
- it is written in rust and seems to have better performance
- it is configured with TOML dotfile instead of ancient profiles
  exporting
- it comes without any native tab an/or pane support which actually is
  good as it enables me to go back to using tmux

Alacritty v1 config TOML is also being added.

I have chosen a catppuccin color theme for it, and this is also a
significant and important change. Catppuccin is a pastel, nice looking
theme, that is available for a lot of tools that I am or I am planning
to be using. Choosing it gives me a consistency across whole toolchain.
I will probably introduce catppuccin as a theme for more things than
Alacritty only.