Add tmux configuration
As alacritty does not do any window, tab or pane management, I'm going
back to using tmux and adding a new config for it.

This is for now mostly a collection of sensible settings, combined with
catpuccin plugin that is configured to show some required stats in the
bar to enable working with alacritty and tmux in a fullscreen mode.