Reinforce tmux settings for neovim
This commit iterates over some small changes for tmux configuration.

I am setting an RGB mode for the terminal emulators in tmux so that it
can properly display colors in applications run inside it that support
it, such as for example neovim.

To make it work with neovim copy buffers I am also adding one option
set as on.

As I am planning to go back to using vim in form of neovim, I am also
moving main tmux bar to the top, as I will have a vim status bar at the
bottom and I would like to have these two status bars separated as much
as possible.