Remove ulimit from zsh config
As I have encountered a problem with running capybara spec battery some
time ago, there was a ulimit option set in main zsh config that was
preventing it from happening.

This commit removes this from main zsh config. In general, solutions for
such local problems should be resolved with local sourcing solutions or
treated as machine centric and therefore placed in profile files
(according to my personal taste, non sharable or secret configs should
happen in profile files such as .zprofile or .bash_profile and should
not be synced to any outside world repo). That ulimit setting is also
too global as there is no reason to block all processed that are run in
the terminal to have any limitations on their number of open file
descriptors.