Add code diagnostics full message
As I stated working with more LSPs, I needed to add a shortcut for
displaying a full lsp diagnostic message in a floating window. Inline
diagnostics are ok for fast checking things, but sometimes I need to
actually read the whole message and I'd prefer to show it in a separate,
easy to close floating window on top of loc.