Add install script
Previously this repo readme was only mentioning some packages and apps
for which config is being provided, that need to be installed for this
configs to make sense and be applied.

I am instroducing singularity.zsh, a script that uses brew and install
not only things for which dotfiles in this repo exist, but also other
apps and packages that I commonly used.

This script is intended to be run during new machine setup, after
installing homebrew.