Do not run alacritty in fullscreen mode
As a preparation step fro trying out custom bars and tiling manager, I
do no longer want to run alacritty terminal in a fullscreen mode.

This commit changes alacritty config so it now starts maximised, and
it's paddings are being modified to better display content when it is
not in fullscreen mode.