Add tmux-powerline
So far, tmux status line was using a catpuccin plugin. Although it was
working correct, it was strange to bind functional parts of tmux status
bar to a color scheme.

This commit replaces tmux catppuccin plugin with a custom styled
tmux-powerline.

Tmux powerline allows to build status bar for tmux with modules, similar
to powerline for zsh or oh-my-posh that we use for prompt. Previous
catpuccin frappe colors are added manually back and I tried to recreate
previous feel and look.

To add nerdfont icons to modules, I needed to recreate base
tmux-powerline segments as custom ones to return strings with icons for
modules that I use.

Using theme agnostic plugin, allows doing multiple color themes for
whole setup in the future, that I would be able to pick from. I can also
go back to the idea of always running alacritty in a fullscreen mode, as
I can now easily use modules such as battery to recreate system status
bar in tmux.