Remove sketchybar
sketchybar turned out to be more maintanance heavy than I wanted it to
be. With a couple of updaes parts of my config stopped working and I
didn't want to fix them. Also, workspaces view was not used by me that
much. Given all of the above, I decided to get rid of sketchybar
completely.

For now I am going back to Mac top bar, maybe with some slight
modifications like Ice to be tried in the future.